// guipmc-casuistica/src/app/casuistica/models/patronDTO.ts
export class PatronDTO {
  fecAlta: Date;
  fecBaja: Date;
  fecActualizacion: Date;
  refRegistroPatronal: string;
  desRazonSocial: string;
  desRfc: string;
  objectIdPatron: string;
  cveClase: string;
  cveDelRegPatronal: string;
  cveFraccion: string;
  numPrima: string;
  cveSubDelRegPatronal: string;
  desClase: string;
  desDelRegPatronal: string;
  desFraccion: string;
  desPrima: string;
  desSubDelRegPatronal: string;
  cveDelegacionAux: string;
  cveModalidad: string;
  actividadEconomica:string;
  delegacionRegistro:string;
}